Tidying up, creating regions in commands/mod.rs
