+++ apollo/src/commands/mod.rs
@@ -1,6 +1,59 @@
-pub mod login;
+use structopt::StructOpt;
+
+/// Commands implement the Command trait, which lets us run() them
+/// and get Output.
+pub trait Command {
+    /// Execute the command. TODO: should this return a Result?
+    fn run(&self) {}
+}
+
+//#region    apollo <command>
+#[derive(StructOpt)]
+#[structopt(rename_all = "kebab-case")]
+/// The [Experimental] Apollo CLI, for supporting all your graphql needs :)
+pub enum Apollo {
+    ///  🖨   parse and pretty print schemas to stdout
+    Print(Print),
+    ///  🔓  log in to apollo
+    Login(Login),
+}
+//#endregion
+
+//#region    ... print [-h] <files...>
 pub mod print;
 
-pub trait Command {
-    fn run(&self) {}
+#[derive(StructOpt)]
+#[structopt(rename_all = "kebab-case")]
+pub struct Print {
+    #[structopt(short = "h", long)]
+    /// suppress headers when printing multiple files
+    pub no_headers: bool,
+
+    #[structopt(parse(from_os_str))]
+    /// schemas to print
+    pub files: std::vec::Vec<std::path::PathBuf>,
+}
+//#endregion
+
+//#region    ... login
+pub mod login;
+
+#[derive(StructOpt)]
+#[structopt(rename_all = "kebab-case")]
+pub struct Login {}
+//#endregion
+
+impl Command for Apollo {
+    fn run(&self) {
+        match self {
+            Apollo::Print(cmd) => cmd.run(),
+            Apollo::Login(cmd) => cmd.run(),    
+        }
+    }
+}
+
+impl Apollo {
+    pub fn main() {
+        Apollo::from_args().run();
+    }
 }
