+++ apollo/src/main.rs
@@ -1,14 +1,6 @@
-use structopt::StructOpt;
-
 mod commands;
-use commands::Command;
-
-mod command_config;
-use command_config::Apollo;
+use commands::Apollo;
 
 fn main() {
-    match Apollo::from_args() {
-        Apollo::Print(cmd) => cmd.run(),
-        Apollo::Login(cmd) => cmd.run(),
-    }
+    Apollo::main()
 }
